+++ globals.py
@@ -1,0 +1,9 @@
+from tkinter import Tk
+import threading
+
+root = Tk()
+lock = threading.Lock()
+gameClosedEvent = threading.Event()
+connClosedEvent = threading.Event()
+myTurn = False
+
