+++ main.py
@@ -4,33 +4,14 @@ from gameBoard import GameBoard
 from gameState import Player, GameState
 from userInterface import UserInterface
 from constants import *
-import threading
 import time
 from clientConnection import *
+from globals import *
 
-connAlive = True
-gameAlive = True
-
-def conn_thread():
-    n = 0
-    while n<20 and gameAlive:
-        print("Hello World: ", n)
-        time.sleep(0.25)
-        n+=1
-
-def checkConn():
-    if threadConn.is_alive():
-        print("Connection alive")
-        root.after(200, checkConn)
-    else:
-        root.destroy()
-
-threadConn = threading.Thread(target=conn_thread)
 if __name__ == "__main__":
-    
-    threadConn.start()
 
-    root = Tk()
+    threadConn.start()
+    #root = Tk()
     window = Window(WINDOW_WIDTH, WINDOW_HEIGHT, root)
     userInterface = UserInterface(root)
     board = GameBoard(window, root, userInterface)
@@ -39,5 +20,7 @@ if __name__ == "__main__":
     gameState = GameState(player1, player2, board, userInterface)
     root.after(200,checkConn)
     root.mainloop()
-    print ("AFter mainloop")
+    lock.acquire()
+    gameClosedEvent.set()
+    lock.release()
     threadConn.join
